requirements.txt
boto3==1.34.0
pyyaml==6.0.2
pytest==8.3.3
moto==5.0.1
flake8==7.1.1